Improved style for toolbar
Active elements are no longer yellow, instead a discreete border is
appended. Additionally there are more space between menu elements.

To achieve proper spacing, it was necessary to slightly change the toolbar
HTML structure. It is now wrapped in a div element. This does not seem to cause
any side-effects.